updated script
